Project Management Part 1

+++ backend/freelancePM/src/main/java/com/freelance/freelancepm/dto/ProjectResponse.java
@@ -2,15 +2,14 @@ package com.freelance.freelancepm.dto;
 
 import lombok.Builder;
 import lombok.Data;
-
 import java.time.LocalDate;
 
 @Data
 @Builder
 public class ProjectResponse {
-    private Long id;
-    private Long clientId;
-    private Long managerId;
+    private Integer id;
+    private Integer clientId;
+    private Integer managerId;
     private String name;
     private String description;
     private String type;
